Add /etc/iftab for persistent network naming
Network interfaces will be named according to their bus info, eth0 should *always* be the control port and eth1 should be the embedded port (in a real instrument). 

eth0 was used to ensure that in a virtual environment (where these rules may fail) the firewall and other such network configurations remain operational.

eth1 may be renamed to a more descriptive name at a later stage. 

Expect an update to the iftab file when the real instrument is available to determine the bus structure.